Added in quote of the day and weather
Weather was already in the code, but now it is email ready. 
QOTD has been added to email, as well.
Comments are now more prevalent.
Took out monkey pox cases for the time being because the websites were changing.
Commented out final block of code because this code is now being run on a server and input is no longer needed.